+++ worker/imap/connect.go
@@ -80,6 +80,11 @@ func (w *IMAPWorker) connect() (*client.Client, error) {
 				username, password, c); err != nil {
 				return nil, err
 			}
+		} else if w.config.xoauth2.Enabled {
+			if err := w.config.xoauth2.Authenticate(
+				username, password, c); err != nil {
+				return nil, err
+			}
 		} else if err := c.Login(username, password); err != nil {
 			return nil, err
 		}